Multiple dbs unique data buckets and unique index buckets simple test
